파일 업로드 예제 수정

+++ ccs-cmn/src/main/java/ccs/cmn/service/impl/FileServiceImpl.java
@@ -1,25 +1,24 @@
 package ccs.cmn.service.impl;
 
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import javax.annotation.Resource;
+
 import org.apache.commons.collections.CollectionUtils;
-import org.springframework.util.ObjectUtils;
+import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
 import ccs.cmn.mapper.FileMapper;
 import ccs.cmn.service.FileService;
-import ccs.cmn.service.UploadFileService;
 import ccs.framework.util.HashMapUtility;
-import lombok.RequiredArgsConstructor;
 
-@RequiredArgsConstructor
+@Service("fileService")
 public class FileServiceImpl implements FileService {
 	
-	private final FileMapper fileMapper;
+	@Resource(name="FileMapper")
+	private FileMapper fileMapper;
 	
 	@Override
 	public List<Map<String, Object>> selectFiles(String FILE_GRUP) throws Exception {
@@ -36,19 +35,22 @@ public class FileServiceImpl implements FileService {
 	public String save(String FILE_GRUP, Map<String, Object> systemParameter, List<Map<String, Object>> filesToInsert,
 			List<Map<String, Object>> filesToDelete) throws Exception {
 		//颇老 昏力
-		filesToDelete.stream().forEach(x->fileMapper.deleteFile(x));
-	
-		//颇老 历厘
-		Map<String, Object> param = HashMapUtility.<String,Object>create().add(systemParameter).toMap();
-		if(StringUtils.isEmpty(FILE_GRUP)) {
-			fileMapper.insertFileGroup(param);
+		if(!CollectionUtils.isEmpty(filesToDelete)) {
+			filesToDelete.stream().forEach(x->fileMapper.deleteFile(x));
 		}
-		final String fileGrupId = FILE_GRUP;
-		filesToInsert.stream().forEach(x->{
-											x.putAll(systemParameter);
-											x.putAll(param);
-											fileMapper.insertFile(x);
-										});
+		//颇老 历厘
+		if(!CollectionUtils.isEmpty(filesToInsert)) {
+			Map<String, Object> param = HashMapUtility.<String,Object>create().add(systemParameter).toMap();
+			if(StringUtils.isEmpty(FILE_GRUP) ) {
+				fileMapper.insertFileGroup(param);
+			}
+			final String fileGrupId = FILE_GRUP;
+			filesToInsert.stream().forEach(x->{
+												x.putAll(systemParameter);
+												x.putAll(param);
+												fileMapper.insertFile(x);
+											});
+		}
 		
 		//return File_group
 		return FILE_GRUP;
